Keep paste fidelity fixes within Swift file budgets
The workflow guard rejected the review-fix growth in two already-large Swift files. The paste fidelity helpers and focused regression tests now live in small dedicated files, preserving the behavior while keeping the budgeted files below their tracked limits.

Constraint: workflow-guard-tests enforces .github/swift-file-length-budget.tsv
Rejected: Refresh the file length budget | the growth is easy to split into focused files
Confidence: high
Scope-risk: narrow
Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv
Tested: ./scripts/reload.sh --tag issue-3069-paste-non-ascii
Not-tested: Local XCTest suite per repo policy; CI will run unit tests
